feat: typesafe mutations and api routes

+++ pages/api/collection.ts
@@ -1,7 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { unstable_getServerSession } from 'next-auth';
+import { z } from 'zod';
 
 import { prisma } from '@/lib/prismadb';
+import { httpCodes, responseMessages } from '@/utils/apiResponses';
 import { infinitePostsCount } from '@/utils/infinitePostsCount';
 import { string } from '@/utils/string';
 import { transformCollectionPost } from '@/utils/transformCollectionPost';
@@ -13,6 +15,20 @@ import { InfinitePosts } from '@/pages/api/post/infinitePosts';
 
 const COLLECTION_POSTS_PER_SCROLL = 9;
 
+const DeleteCollectionSchema = z.object({
+  postId: z.string(),
+});
+
+export const PutCollectionSchema = z.object({
+  userId: z.string(),
+  postId: z.string(),
+});
+
+const GetCollectionSchema = z.object({
+  skip: z.string(),
+  userId: z.string(),
+});
+
 const apiHandler = async (req: NextApiRequest, res: NextApiResponse) => {
   const { method } = req;
   const session = await unstable_getServerSession(req, res, authOptions);
@@ -22,49 +38,72 @@ const apiHandler = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 
   if (method === 'DELETE') {
+    const response = DeleteCollectionSchema.safeParse(req.query);
+
+    if (!response.success) {
+      return res.status(httpCodes.badRequest).send({
+        message: responseMessages.badRequest,
+      });
+    }
+
     try {
-      const { postID } = req.query;
+      const { postId } = response.data;
       await prisma.collection.deleteMany({
         where: {
-          post_id: Number(string(postID)),
+          post_id: Number(string(postId)),
           user_id: session.user?.id,
         },
       });
 
-      res.status(200).send({ status: 'resource updated successfully' });
+      res.status(httpCodes.resourceSuccess).send(responseMessages.resourceSuccess);
     } catch (error) {
-      res.status(400).send({ status: 'error', message: 'Cannot remove post from your collection' });
+      res.status(httpCodes.badRequest).send(responseMessages.badRequest);
     }
   }
 
   if (method === 'PUT') {
+    const response = PutCollectionSchema.safeParse(req.body);
+
+    if (!response.success) {
+      return res.status(httpCodes.badRequest).send({
+        message: responseMessages.badRequest,
+      });
+    }
     try {
-      const { userID, postID } = req.body;
+      const { userId, postId } = response.data;
+
       await prisma.collection.create({
         data: {
-          post_id: postID,
-          user_id: userID,
+          post_id: parseInt(postId),
+          user_id: userId,
         },
       });
 
-      res.status(201).send({ status: 'resource updated successfully' });
+      res.status(httpCodes.resourceSuccess).send(responseMessages.resourceSuccess);
     } catch (error) {
-      res.status(400).send({ status: 'error', message: 'Cannot add post to your collection' });
+      res.status(httpCodes.badRequest).send(responseMessages.badRequest);
     }
   }
 
   if (method === 'GET') {
-    const { skip, user } = req.query;
+    const response = GetCollectionSchema.safeParse(req.query);
+
+    if (!response.success) {
+      return res.status(httpCodes.badRequest).send({
+        message: responseMessages.badRequest,
+      });
+    }
+
+    const { skip, userId } = response.data;
     const skipNumber = parseInt(string(skip));
-    const takeNumber = COLLECTION_POSTS_PER_SCROLL;
 
     try {
       const posts = await prisma.collection.findMany({
-        skip: skipNumber * takeNumber,
-        take: takeNumber,
+        skip: skipNumber * COLLECTION_POSTS_PER_SCROLL,
+        take: COLLECTION_POSTS_PER_SCROLL,
 
         where: {
-          user_id: string(user),
+          user_id: userId,
         },
 
         include: {
@@ -107,9 +146,9 @@ const apiHandler = async (req: NextApiRequest, res: NextApiResponse) => {
         cursor: nextCursor,
       };
 
-      res.status(200).send(response);
+      res.status(httpCodes.success).send(response);
     } catch (error) {
-      res.status(400).send({ status: 'error', message: 'Cannot download your collection' });
+      res.status(httpCodes.badRequest).send(responseMessages.badRequest);
     }
   }
 };
